Changed all the delete link button to delete form button

+++ view/public/delete-employee.php
@@ -15,17 +15,19 @@ require('../include/html_head.php');
     <main>
         <?php
 
-        $id = $_GET['id'];
+        if (!empty($_POST['id'])) {
+            $id = $_POST['id'];
 
-        $sql = "DELETE FROM employees where id=$id";
-        $ret = $db->exec($sql);
+            $sql = "DELETE FROM employees where id=$id";
+            $ret = $db->exec($sql);
 
-        if (!$ret) {
-            echo $db->lastErrorMsg();
-        } else {
-            echo '<div class="alert alert-success" role="alert">';
-            echo $db->changes(), " Rcord Deleted Successfully..!";
-            echo '</div>';
+            if (!$ret) {
+                echo $db->lastErrorMsg();
+            } else {
+                echo '<div class="alert alert-success" role="alert">';
+                echo $db->changes(), " Rcord Deleted Successfully..!";
+                echo '</div>';
+            }
         }
 
         ?>
